requirements.txt
streamlit
plotly
pydeck
Pillow
vaderSentiment